Update
- Added favicon
- Added follows
- Added favorites
- Added CSRF tokens
- Added error page
- Added 'all uploads' page
- Added follows page
- Added favorites page
- Added users search function
- Added view counts
- Added default profile picture

+++ edit-comment.php
@@ -23,28 +23,32 @@ if(!$id = Input::get('id')) {
 $page_title = "Upload - Edit Comment";
 
 if(Input::exists($_POST, 'edit_comment')) {
-  $validate = new Validate();
+  if(Hash::checkToken(Input::get('token'), 'token')) {
+    $validate = new Validate();
 
-  $validation = $validate->check($_POST, array(
-    'comment_text' => array(
-      'required' => true,
-      'min' => 1,
-      'max' => 255
-    )
-  ));
-
-  if($validation->passed()) {
-    if($upload->editComment($id, array('comment_text' => escape(Input::get('comment_text'))))) {
-      Redirect::to(BASE_URL . '/view?id=' . $comment->comment_upload);
-    } else {
-      $validation->addError("Unable to update comment");
+    $validation = $validate->check($_POST, array(
+      'comment_text' => array(
+        'required' => true,
+        'min' => 1,
+        'max' => 255
+      )
+    ));
+  
+    if($validation->passed()) {
+      if($upload->editComment($id, array('comment_text' => escape(Input::get('comment_text'))))) {
+        Redirect::to(BASE_URL . '/view?id=' . $comment->comment_upload);
+      } else {
+        $validation->addError("Unable to update comment");
+      }
     }
   }
 }
 
 if(Input::exists($_POST, 'delete_comment')) {
-  if($upload->deleteComment($id)) {
-    Redirect::to(BASE_URL . '/view?id=' . $comment->comment_upload);
+  if(Hash::checkToken(Input::get('delete-token'), 'delete-token')) {
+    if($upload->deleteComment($id)) {
+      Redirect::to(BASE_URL . '/view?id=' . $comment->comment_upload);
+    }
   }
 }
